Improve type safety. (#2)
* Require setting Component.type.

Instead of asking subclasses to set the type variable, Component should require them to pass it as an argument. That way, if you forget, Haxe will remind you.

* Make getComponent() type-safe.

As @nadako suggested on Twitter.

* Fix breaking changes

This time, it uses reflection to find the class, meaning you don't have to pass in an extra parameter.

Also, it looks like the generic isn't actually required, so I got rid of it. That way, you don't have to type out `Component<ComponentSubclass>` or `Component<Dynamic>` every time.

* Add casts

This part of the code cannot be made type safe. At least, not without seriously inconveniencing the user by forcing them to type out `Component<ComponentSubclass>` every time.

The important type checks are done elsewhere, anyway. (`c.type` must be a `Class<Component>`, and `classType` must be a `Class<SomethingThatExtendsComponent>`.)

* Make removeComponent() generic as well

Like with getComponent(), this will improve code completion.